distribution-types: small IndexUrls refactor
This removes all public methods on `IndexUrls` except for `indexes`. The
`index` and `extra_index` methods weren't being used anywhere else, so
those were just un-exported. The `no_index` method was being used to
check if the `--no-index` parameter was provided, but we can actually
encapsulate that by simply not returning any indexes when it's set.

I did this so that callers don't need to know whether `no_index` is
enabled or not. They just check the iterator given to them instead.